+++ lib/enschin/include/enschin/game.h
@@ -23,21 +23,14 @@ private:
     bool running = false;
     std::string gameName;
 public:
-    Game(std::string gameName, Vec2f windowSize, bool fullscreen);
+    Game(std::string gameName, Vec2i windowSize, bool fullscreen);
 
     void init();
-
     virtual void start(const GameContext &gtx) = 0;
-
     virtual void loop(const GameContext &gtx) = 0;
-
     virtual void free();
-
     float getDeltaTime() const { return deltaTime; }
-
     float getTotalTime() const { return totalTime; }
-
     float getFps() { return fps; }
-
     Window &getWindow() { return window; }
 };
