+++ lib/enschin/include/enschin/sprite_sheet.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "vec2f.h"
+#include "vec2i.h"
 #include "texture.h"
 
 struct SpriteSheet {
@@ -11,14 +11,10 @@ private:
     unsigned char fps;
 public:
     SpriteSheet() = default;
-
-    SpriteSheet(const std::string &filePath, Vec2f spriteSize, unsigned char fps);
-
+    SpriteSheet(const std::string &filePath, Vec2i spriteSize, unsigned char fps);
     ~SpriteSheet();
 
     Texture *getTexture(int index = 0) const { return textures[index]; }
-
     unsigned short getAmountOfSprites() { return amountOfSprites; }
-
     unsigned char getFps() { return fps; }
 };
