+++ lib/enschin/include/enschin/window.h
@@ -6,7 +6,7 @@
 #include <string>
 #include <iostream>
 #include "../../deps/stb/stb_image.h"
-#include "vec2f.h"
+#include "vec2i.h"
 #include "color.h"
 #include "renderer.h"
 
@@ -15,32 +15,22 @@
  */
 class Window {
 private:
-    Vec2f windowSize;
+    Vec2i windowSize;
     bool fullScreen;
     bool vSync = true;
     GLFWwindow *window;
 public:
     Window() = default;
-
-    Window(std::string title, Vec2f size, bool fullscreen);
+    Window(std::string title, Vec2i size, bool fullscreen);
 
     bool update();
-
-    Vec2f getSize() { return windowSize; }
-
+    Vec2i getSize() { return windowSize; }
     GLFWwindow *getGlfw() { return window; }
-
-    void setSize(Vec2f v);
-
+    void setSize(Vec2i v);
     void setFullScreen(bool fullscreen);
-
     void setTitle(std::string title);
-
     void setClearColor(Color newColor);
-
     void setVSync(bool vSync);
-
     void setIcon(const char *filePath);
-
     void setAntialiasing(bool state) { if (state)glEnable(GL_MULTISAMPLE); else glDisable(GL_MULTISAMPLE); }
 };
